no pub field but to_jiff and to_sqlx
